feat: rename `pomodoroFluentWindow.py` to `koncentroFluentWindow.py`
